refactor(highlight): isolate sync tree-sitter logic in HighlightSyncService
