+++ crates/mcb-infrastructure/src/services/highlight_service.rs
@@ -18,280 +18,29 @@ use std::sync::Arc;
 
 use mcb_domain::ports::{HighlightError, HighlightServiceInterface};
 use mcb_domain::registry::services::ServiceBuilder;
-use mcb_domain::value_objects::browse::{HighlightSpan, HighlightedCode};
-use tree_sitter::Language;
-use tree_sitter_highlight::{Highlight, HighlightConfiguration, HighlightEvent, Highlighter};
+use mcb_domain::value_objects::browse::HighlightedCode;
 
-use mcb_domain::value_objects::browse::{HIGHLIGHT_NAMES, map_highlight_to_category};
-
-/// Language-specific highlighting configuration
-struct HighlightLanguageConfig {
-    name: &'static str,
-    language: Language,
-    highlights_query: &'static str,
-}
-
-impl HighlightLanguageConfig {
-    fn new(name: &'static str, language: Language, highlights_query: &'static str) -> Self {
-        Self {
-            name,
-            language,
-            highlights_query,
-        }
-    }
-}
-
-/// Internal state holding both the highlighter and cached language configs.
-struct HighlighterState {
-    highlighter: Highlighter,
-    configs: std::collections::HashMap<String, HighlightConfiguration>,
-}
+use crate::services::highlight_sync_service::{HighlightSyncPort, HighlightSyncService};
 
 /// Concrete highlight service implementation using tree-sitter.
 ///
-/// Manages a thread-safe `Highlighter` instance and **cached** language configurations
-/// to perform efficient, on-demand syntax highlighting.
+/// Delegates all CPU-bound work to [`HighlightSyncService`] running inside
+/// `spawn_blocking` so the async executor is never blocked.
 pub struct HighlightServiceImpl {
-    state: Arc<std::sync::Mutex<HighlighterState>>,
+    inner: Arc<dyn HighlightSyncPort>,
 }
 
 impl HighlightServiceImpl {
-    /// Creates a syntax highlight service with an internal tree-sitter highlighter.
+    /// Creates a syntax highlight service wrapping the given sync port.
     #[must_use]
-    pub fn new() -> Self {
-        Self {
-            state: Arc::new(std::sync::Mutex::new(HighlighterState {
-                highlighter: Highlighter::new(),
-                configs: std::collections::HashMap::new(),
-            })),
-        }
-    }
-
-    fn get_language_config_dynamic(
-        lang_id: mcb_domain::ports::validation::LanguageId,
-    ) -> Option<Result<HighlightLanguageConfig, HighlightError>> {
-        #[allow(clippy::wildcard_enum_match_arm)]
-        match lang_id {
-            mcb_domain::ports::validation::LanguageId::Python => {
-                Some(Ok(HighlightLanguageConfig::new(
-                    "python",
-                    tree_sitter_python::LANGUAGE.into(),
-                    tree_sitter_python::HIGHLIGHTS_QUERY,
-                )))
-            }
-            mcb_domain::ports::validation::LanguageId::JavaScript => {
-                Some(Ok(HighlightLanguageConfig::new(
-                    "javascript",
-                    tree_sitter_javascript::LANGUAGE.into(),
-                    tree_sitter_javascript::HIGHLIGHT_QUERY,
-                )))
-            }
-            mcb_domain::ports::validation::LanguageId::TypeScript => {
-                Some(Ok(HighlightLanguageConfig::new(
-                    "typescript",
-                    tree_sitter_typescript::LANGUAGE_TYPESCRIPT.into(),
-                    tree_sitter_typescript::HIGHLIGHTS_QUERY,
-                )))
-            }
-            mcb_domain::ports::validation::LanguageId::Tsx => {
-                Some(Ok(HighlightLanguageConfig::new(
-                    "tsx",
-                    tree_sitter_typescript::LANGUAGE_TSX.into(),
-                    tree_sitter_typescript::HIGHLIGHTS_QUERY,
-                )))
-            }
-            mcb_domain::ports::validation::LanguageId::Ruby => {
-                Some(Ok(HighlightLanguageConfig::new(
-                    "ruby",
-                    tree_sitter_ruby::LANGUAGE.into(),
-                    tree_sitter_ruby::HIGHLIGHTS_QUERY,
-                )))
-            }
-            mcb_domain::ports::validation::LanguageId::Php => {
-                Some(Ok(HighlightLanguageConfig::new(
-                    "php",
-                    tree_sitter_php::LANGUAGE_PHP.into(),
-                    tree_sitter_php::HIGHLIGHTS_QUERY,
-                )))
-            }
-            _unsupported_lang => None,
-        }
-    }
-
-    fn get_language_config_static(
-        lang_id: mcb_domain::ports::validation::LanguageId,
-    ) -> Option<Result<HighlightLanguageConfig, HighlightError>> {
-        #[allow(clippy::wildcard_enum_match_arm)]
-        match lang_id {
-            mcb_domain::ports::validation::LanguageId::Rust => {
-                Some(Ok(HighlightLanguageConfig::new(
-                    "rust",
-                    tree_sitter_rust::LANGUAGE.into(),
-                    tree_sitter_rust::HIGHLIGHTS_QUERY,
-                )))
-            }
-            mcb_domain::ports::validation::LanguageId::Go => {
-                Some(Ok(HighlightLanguageConfig::new(
-                    "go",
-                    tree_sitter_go::LANGUAGE.into(),
-                    tree_sitter_go::HIGHLIGHTS_QUERY,
-                )))
-            }
-            mcb_domain::ports::validation::LanguageId::Java => {
-                Some(Ok(HighlightLanguageConfig::new(
-                    "java",
-                    tree_sitter_java::LANGUAGE.into(),
-                    tree_sitter_java::HIGHLIGHTS_QUERY,
-                )))
-            }
-            mcb_domain::ports::validation::LanguageId::C => Some(Ok(HighlightLanguageConfig::new(
-                "c",
-                tree_sitter_c::LANGUAGE.into(),
-                tree_sitter_c::HIGHLIGHT_QUERY,
-            ))),
-            mcb_domain::ports::validation::LanguageId::Cpp => {
-                Some(Ok(HighlightLanguageConfig::new(
-                    "cpp",
-                    tree_sitter_cpp::LANGUAGE.into(),
-                    tree_sitter_cpp::HIGHLIGHT_QUERY,
-                )))
-            }
-            mcb_domain::ports::validation::LanguageId::Swift => {
-                Some(Ok(HighlightLanguageConfig::new(
-                    "swift",
-                    tree_sitter_swift::LANGUAGE.into(),
-                    tree_sitter_swift::HIGHLIGHTS_QUERY,
-                )))
-            }
-            _unsupported_lang => None,
-        }
-    }
-
-    /// Get language configuration for supported languages
-    fn get_language_config(language: &str) -> Result<HighlightLanguageConfig, HighlightError> {
-        let lang_id = mcb_domain::ports::validation::LanguageId::from_name(language)
-            .ok_or_else(|| HighlightError::UnsupportedLanguage(language.to_owned()))?;
-
-        if let Some(res) = Self::get_language_config_dynamic(lang_id) {
-            return res;
-        }
-
-        if let Some(res) = Self::get_language_config_static(lang_id) {
-            return res;
-        }
-
-        Err(HighlightError::UnsupportedLanguage(language.to_owned()))
-    }
-
-    /// Create highlight configuration from language config
-    fn create_highlight_config(
-        lang_config: HighlightLanguageConfig,
-    ) -> Result<HighlightConfiguration, HighlightError> {
-        let mut config = HighlightConfiguration::new(
-            lang_config.language,
-            lang_config.name,
-            lang_config.highlights_query,
-            "",
-            "",
-        )
-        .map_err(|e| HighlightError::ConfigurationError(e.to_string()))?;
-
-        config.configure(&HIGHLIGHT_NAMES);
-        Ok(config)
-    }
-
-    fn parse_highlight_events(
-        highlights: impl Iterator<Item = Result<HighlightEvent, tree_sitter_highlight::Error>>,
-    ) -> Result<Vec<HighlightSpan>, HighlightError> {
-        let mut spans = Vec::new();
-        let mut position = 0;
-        let mut open_spans: Vec<(usize, &str)> = Vec::new();
-
-        for event in highlights {
-            match event {
-                Ok(HighlightEvent::Source { end, .. }) => {
-                    position = end;
-                }
-                Ok(HighlightEvent::HighlightStart(Highlight(highlight))) => {
-                    if let Some(class_name) = HIGHLIGHT_NAMES.get(highlight) {
-                        open_spans.push((position, class_name));
-                    }
-                }
-                Ok(HighlightEvent::HighlightEnd) => {
-                    if let Some((start, class_name)) = open_spans.pop() {
-                        let category = map_highlight_to_category(class_name);
-                        spans.push(HighlightSpan {
-                            start,
-                            end: position,
-                            category,
-                        });
-                    }
-                }
-                Err(e) => {
-                    return Err(HighlightError::HighlightingFailed(e.to_string()));
-                }
-            }
-        }
-
-        Ok(spans)
-    }
-
-    /// Highlight code using tree-sitter with cached language configs.
-    fn highlight_code_internal(
-        &self,
-        code: &str,
-        language: &str,
-    ) -> Result<HighlightedCode, HighlightError> {
-        if code.is_empty() {
-            return Ok(HighlightedCode {
-                original: code.to_owned(),
-                spans: vec![],
-                language: language.to_owned(),
-            });
-        }
-
-        let mut state = self
-            .state
-            .lock()
-            .map_err(|e| HighlightError::HighlightingFailed(format!("Lock poisoned: {e}")))?;
-
-        // Destructure to allow independent borrows of configs and highlighter
-        let HighlighterState {
-            highlighter,
-            configs,
-        } = &mut *state;
-
-        // Get or create cached config for this language (entry API avoids expect)
-        let config = match configs.entry(language.to_owned()) {
-            std::collections::hash_map::Entry::Occupied(e) => e.into_mut(),
-            std::collections::hash_map::Entry::Vacant(e) => {
-                let lang_config = Self::get_language_config(language)?;
-                let config = Self::create_highlight_config(lang_config).map_err(|e| {
-                    HighlightError::ConfigurationError(format!(
-                        "failed to create highlight config for '{language}': {e}"
-                    ))
-                })?;
-                e.insert(config)
-            }
-        };
-
-        let highlights = highlighter
-            .highlight(config, code.as_bytes(), None, |_: &str| None)
-            .map_err(|e| HighlightError::HighlightingFailed(e.to_string()))?;
-
-        let spans = Self::parse_highlight_events(highlights)?;
-
-        Ok(HighlightedCode {
-            original: code.to_owned(),
-            spans,
-            language: language.to_owned(),
-        })
+    pub fn new(inner: Arc<dyn HighlightSyncPort>) -> Self {
+        Self { inner }
     }
 }
 
 impl Default for HighlightServiceImpl {
     fn default() -> Self {
-        Self::new()
+        Self::new(Arc::new(HighlightSyncService::new()))
     }
 }
 
@@ -300,14 +49,13 @@ impl HighlightServiceInterface for HighlightServiceImpl {
     async fn highlight(&self, code: &str, language: &str) -> mcb_domain::Result<HighlightedCode> {
         let code = code.to_owned();
         let language = language.to_owned();
-        let state = Arc::clone(&self.state);
+        let inner = Arc::clone(&self.inner);
 
-        let result = tokio::task::spawn_blocking(move || {
-            let service = HighlightServiceImpl { state };
-            service.highlight_code_internal(&code, &language)
-        })
-        .await
-        .map_err(|e| HighlightError::HighlightingFailed(format!("Blocking task failed: {e}")))?;
+        let result = tokio::task::spawn_blocking(move || inner.highlight(&code, &language))
+            .await
+            .map_err(|e| {
+                HighlightError::HighlightingFailed(format!("Blocking task failed: {e}"))
+            })?;
 
         result.map_err(mcb_domain::Error::from)
     }
@@ -315,5 +63,9 @@ impl HighlightServiceInterface for HighlightServiceImpl {
 
 mcb_domain::register_service!(
     mcb_utils::constants::SERVICE_NAME_HIGHLIGHT,
-    ServiceBuilder::Highlight(|_context| { Ok(std::sync::Arc::new(HighlightServiceImpl::new())) }),
+    ServiceBuilder::Highlight(|_context| {
+        Ok(std::sync::Arc::new(HighlightServiceImpl::new(Arc::new(
+            HighlightSyncService::new(),
+        ))))
+    }),
 );
